+++ internal/chat/person.go
@@ -2,6 +2,6 @@ package chat
 
 type Person struct {
 	ID             uint64
-	TelegramUserID int
+	TelegramUserID int64
 	TelegramChatID int64
 }
